move work related to Controller concept in a new module (pippo-controller)
