+++ pippo-controller/pom.xml
@@ -9,28 +9,16 @@
 
     <modelVersion>4.0.0</modelVersion>
     <packaging>jar</packaging>
-    <artifactId>pippo-ioc</artifactId>
+    <artifactId>pippo-controller</artifactId>
     <version>0.4.0-SNAPSHOT</version>
-    <name>Pippo IoC</name>
-    <description>Common IoC for injection</description>
+    <name>Pippo Controller</name>
+    <description>Pippo Controller</description>
 
     <dependencies>
         <dependency>
             <groupId>ro.pippo</groupId>
             <artifactId>pippo-core</artifactId>
             <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>ro.pippo</groupId>
-            <artifactId>pippo-controller</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>javax.inject</groupId>
-            <artifactId>javax.inject</artifactId>
-            <version>1</version>
         </dependency>
 	</dependencies>
