+++ pippo-controller/src/main/java/ro/pippo/controller/Body.java
@@ -13,26 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ro.pippo.core.controller;
+package ro.pippo.controller;
 
-import ro.pippo.core.util.ListenerList;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * @author Decebal Suiu
+ * Annotation that identifies that a request body should be mapped to a Java object.
+ *
+ * @author James Moger
  */
-public class ControllerInstantiationListenerList extends ListenerList<ControllerInstantiationListener>
-    implements ControllerInstantiationListener {
-
-    @Override
-    public void onInstantiation(final Controller controller) {
-        notify(new Notifier<ControllerInstantiationListener>() {
-
-            @Override
-            public void notify(ControllerInstantiationListener listener) {
-                listener.onInstantiation(controller);
-            }
-
-        });
-    }
-
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.PARAMETER})
+public @interface Body {
 }
