+++ pippo-demo/pippo-demo-controller/pom.xml
@@ -29,6 +29,12 @@
 
         <dependency>
             <groupId>ro.pippo</groupId>
+            <artifactId>pippo-controller</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>ro.pippo</groupId>
             <artifactId>pippo-freemarker</artifactId>
             <version>${project.version}</version>
         </dependency>
