+++ pippo-demo/pippo-demo-crudng/pom.xml
@@ -29,6 +29,12 @@
 
         <dependency>
             <groupId>ro.pippo</groupId>
+            <artifactId>pippo-controller</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>ro.pippo</groupId>
             <artifactId>pippo-freemarker</artifactId>
             <version>${project.version}</version>
         </dependency>
