+++ pippo-demo/pippo-demo-crudng/src/main/java/ro/pippo/demo/crudng/CrudNgApiController.java
@@ -18,8 +18,8 @@ package ro.pippo.demo.crudng;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ro.pippo.core.Param;
-import ro.pippo.core.controller.Body;
-import ro.pippo.core.controller.Controller;
+import ro.pippo.controller.Body;
+import ro.pippo.controller.Controller;
 import ro.pippo.demo.common.Contact;
 import ro.pippo.demo.common.ContactService;
 import ro.pippo.metrics.Metered;
