+++ pippo-demo/pippo-demo-guice/src/main/java/ro/pippo/demo/guice/ContactsController2.java
@@ -15,7 +15,7 @@
  */
 package ro.pippo.demo.guice;
 
-import ro.pippo.core.controller.Controller;
+import ro.pippo.controller.Controller;
 import ro.pippo.demo.common.ContactService;
 
 /**
